Add security for note endpoints

+++ src/main/java/pl/maciejdudek/project/controllers/NoteController.java
@@ -3,12 +3,16 @@ package pl.maciejdudek.project.controllers;
 import lombok.RequiredArgsConstructor;
 import org.modelmapper.ModelMapper;
 import org.springframework.data.domain.Sort;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
+import pl.maciejdudek.project.configuration.security.CheckPermissions;
+import pl.maciejdudek.project.exceptions.UnauthorizedException;
 import pl.maciejdudek.project.model.DTO.NoteDTO;
 import pl.maciejdudek.project.model.Note;
 import pl.maciejdudek.project.model.NoteStatus;
 import pl.maciejdudek.project.services.NoteServiceImpl;
 
+import java.security.Principal;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
@@ -20,32 +24,44 @@ public class NoteController {
 
     private final NoteServiceImpl noteService;
     private final ModelMapper modelMapper;
+    private final CheckPermissions checkPermissions;
 
-    // todo: add security (only for admin)
+    // only for admin
     @GetMapping("/notes")
     public List<NoteDTO> getAll(@RequestParam(defaultValue = "0") int page, @RequestParam(defaultValue = "5") int size,
-                                @RequestParam(defaultValue = "ASC") Sort.Direction sort, @RequestParam(defaultValue = "id") String by) {
-        return noteService.getAll(page, size, sort, by).stream()
-                .map(note -> modelMapper.map(note, NoteDTO.class))
-                .collect(Collectors.toList());
+                                @RequestParam(defaultValue = "ASC") Sort.Direction sort, @RequestParam(defaultValue = "id") String by,
+                                @AuthenticationPrincipal Principal principal) {
+        if(checkPermissions.isAdmin(principal.getName())) {
+            return noteService.getAll(page, size, sort, by).stream()
+                    .map(note -> modelMapper.map(note, NoteDTO.class))
+                    .collect(Collectors.toList());
+        }
+        throw new UnauthorizedException();
     }
 
-    // todo: add security (only for admin and user who created note)
+    // only for admin and user who created note
     @GetMapping("/notes/{id}")
-    public NoteDTO getOne(@PathVariable Long id) {
-        return modelMapper.map(
-                noteService.getOne(id),
-                NoteDTO.class);
+    public NoteDTO getOne(@PathVariable Long id, @AuthenticationPrincipal Principal principal) {
+        if(checkPermissions.isAdmin(principal.getName()) || checkPermissions.noteBelongsToUser(principal.getName(), id)) {
+            return modelMapper.map(
+                    noteService.getOne(id),
+                    NoteDTO.class);
+        }
+        throw new UnauthorizedException();
     }
 
-    // todo: add security (only for admin and user where userId equals id from request)
+    // only for admin and user where userId equals id from request
     @GetMapping("/users/{id}/notes")
     public List<NoteDTO> getAllByUser(@PathVariable Long id,
                                       @RequestParam(defaultValue = "0") int page, @RequestParam(defaultValue = "5") int size,
-                                      @RequestParam(defaultValue = "ASC") Sort.Direction sort, @RequestParam(defaultValue = "id") String by) {
-        return noteService.getAllByUser(id, page, size, sort, by).stream()
-                .map(note -> modelMapper.map(note, NoteDTO.class))
-                .collect(Collectors.toList());
+                                      @RequestParam(defaultValue = "ASC") Sort.Direction sort, @RequestParam(defaultValue = "id") String by,
+                                      @AuthenticationPrincipal Principal principal) {
+        if(checkPermissions.isAdmin(principal.getName()) || checkPermissions.matchingId(principal.getName(), id)) {
+            return noteService.getAllByUser(id, page, size, sort, by).stream()
+                    .map(note -> modelMapper.map(note, NoteDTO.class))
+                    .collect(Collectors.toList());
+        }
+        throw new UnauthorizedException();
     }
 
     // todo: set default user id (user who creating note)
@@ -56,26 +72,38 @@ public class NoteController {
                 NoteDTO.class);
     }
 
-    // todo: add security (only for user who created note)
+    // only for user who created note
     @PutMapping("/notes/{id}")
-    public NoteDTO update(@PathVariable Long id, @RequestBody NoteDTO noteDTO) {
-        return modelMapper.map(
-                noteService.update(id, modelMapper.map(noteDTO, Note.class)),
-                NoteDTO.class);
+    public NoteDTO update(@PathVariable Long id, @RequestBody NoteDTO noteDTO,
+                          @AuthenticationPrincipal Principal principal) {
+        if(checkPermissions.noteBelongsToUser(principal.getName(), id)) {
+            return modelMapper.map(
+                    noteService.update(id, modelMapper.map(noteDTO, Note.class)),
+                    NoteDTO.class);
+        }
+        throw new UnauthorizedException();
     }
 
-    // todo: add security (only for user who created note)
+    // only for user who created note
     @PutMapping("/notes/{id}/status")
-    public NoteDTO updateStatus(@PathVariable Long id, @RequestBody Map<String, NoteStatus> statusJson) {
-        return modelMapper.map(
-                noteService.updateStatus(id, statusJson.get("status")),
-                NoteDTO.class
-        );
+    public NoteDTO updateStatus(@PathVariable Long id, @RequestBody Map<String, NoteStatus> statusJson,
+                                @AuthenticationPrincipal Principal principal) {
+        if(checkPermissions.noteBelongsToUser(principal.getName(), id)) {
+            return modelMapper.map(
+                    noteService.updateStatus(id, statusJson.get("status")),
+                    NoteDTO.class
+            );
+        }
+        throw new UnauthorizedException();
     }
 
-    // todo: add security (only for admin)
+    // only for admin
     @DeleteMapping("/notes/{id}")
-    public void delete(@PathVariable Long id) {
-        noteService.delete(id);
+    public void delete(@PathVariable Long id, @AuthenticationPrincipal Principal principal) {
+        if(checkPermissions.isAdmin(principal.getName())) {
+            noteService.delete(id);
+        } else {
+            throw new UnauthorizedException();
+        }
     }
 }
